ComputerVisionResearch/version2/yolo_dataset_manager/requirements.txt
streamlit
pyyaml
pandas
matplotlib
opencv-python-headless
Pillow
torch
ultralytics
streamlit-drawable-canvas